requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
python-geohash
python-dotenv
websockets
geopy